pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cs-trading"
version = "0.1.0"
description = "Counter-Strike trading scanner with BUFF market data"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "CS Trading" }]
dependencies = [
  "requests>=2.31.0",
  "pydantic>=2.6.0",
  "python-dateutil>=2.9.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
cs-trading = "cs_trading.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
